[#22] Restructuring common domain
