+++ src/common/infrastructure/service/CommonErrorServiceImpl.ts
@@ -1,5 +1,5 @@
 import RepositoryError, { RepositoryErrorData } from "src/common/domain/RepositoryError";
-import CommonErrorService from "src/common/service/CommonErrorService";
+import CommonErrorService from "src/common/domain/service/CommonErrorService";
 
 class CommonErrorServiceImpl implements CommonErrorService {
   public createRepositoryErrorFrom(e?: RepositoryErrorData): RepositoryError {
