+++ src/common/presentation/components/atmos/ImmutableTextField.tsx
@@ -3,7 +3,7 @@ import { TextFieldProps } from '@material-ui/core/TextField';
 import { makeStyles } from '@material-ui/styles';
 import clsx from 'clsx'
 import * as React from 'react';
-import I18NService from 'src/common/service/I18NService';
+import I18NService from 'src/common/domain/service/I18NService';
 
 const useStyles = makeStyles(createStyles({
   root: {
