+++ src/common/presentation/components/molecules/Confirm.tsx
@@ -4,7 +4,7 @@ import { fade } from '@material-ui/core/styles';
 import { createStyles, makeStyles } from '@material-ui/styles';
 import clsx from 'clsx';
 import * as React from 'react';
-import I18NService from 'src/common/service/I18NService';
+import I18NService from 'src/common/domain/service/I18NService';
 import { ConfirmPayload } from '../../state-module/common';
 
 const useStyles = makeStyles((theme: Theme) => createStyles({
