+++ src/common/presentation/components/templates/CmsLayout.tsx
@@ -18,7 +18,7 @@ import React, { Fragment } from 'react';
 import { WithTranslation } from 'react-i18next';
 import { FirstDepthPath } from 'src/common/domain/constants/FIRST_DEPTH_PATHS';
 import SIDE_BAR_ITEMS from 'src/common/domain/constants/SIDE_BAR_ITEMS';
-import I18NService from 'src/common/service/I18NService';
+import I18NService from 'src/common/domain/service/I18NService';
 import HorizontalMenuBarContainer from '../../container/molecules/HorizontalMenuBarContainer';
 import LanguageToggleButton from '../atmos/LanguageToggleButton';
 import Link from '../atmos/Link';
