--- src/mother/notice/domain/NoticeRepository.ts
@@ -1,4 +1,0 @@
-import CrudRepository from 'src/common/domain/CrudRepository';
-import Notice from './Notice'
-
-export default interface NoticeRepository extends CrudRepository<Notice> { }
