+++ src/mother/notice/domain/model/Notice.ts
