+++ src/mother/notice/domain/repository/NoticeRepository.ts
@@ -1,0 +1,4 @@
+import CrudRepository from 'src/common/domain/repository/CrudRepository';
+import Notice from '../model/Notice'
+
+export default interface NoticeRepository extends CrudRepository<Notice> { }
