+++ src/mother/notice/domain/service/NoticeService.ts
@@ -1,8 +1,8 @@
 import Id from 'src/common/domain/Id';
 import Page from "src/common/domain/Page";
 import PageRequest from "src/common/domain/PageRequest";
-import NoticeRequestDto from "../api/dto/NoticeRequestDto";
-import Notice from "../domain/Notice";
+import NoticeRequestDto from "../../api/dto/NoticeRequestDto";
+import Notice from "../model/Notice";
 
 export default interface NoticeService {
   getNotice(id: Id): Promise<Notice>
