+++ src/mother/notice/presentation/components/organisms/NoticeDetail.tsx
@@ -3,12 +3,12 @@ import { Delete, Edit } from '@material-ui/icons';
 import { createStyles, makeStyles } from '@material-ui/styles';
 import clsx from 'clsx';
 import * as React from 'react';
+import I18NService from 'src/common/domain/service/I18NService';
 import ImmutableTextField from 'src/common/presentation/components/atmos/ImmutableTextField';
 import MarkdownPreview from 'src/common/presentation/components/atmos/previews/MarkdownPreview';
 import ErrorTypography from 'src/common/presentation/components/atmos/typographies/ErrorTypography';
 import MySpeedDial, { SpeedDialActionData } from 'src/common/presentation/components/molecules/MySpeedDial';
-import I18NService from 'src/common/service/I18NService';
-import Notice from 'src/mother/notice/domain/Notice';
+import Notice from 'src/mother/notice/domain/model/Notice';
 import { createLinkClickHandler } from 'src/util/createLinkClickHandler';
 
 const useStyles = makeStyles((theme: Theme) => createStyles({
