+++ src/mother/notice/presentation/containers/NoticeListContainer.tsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { bindActionCreators, Dispatch } from 'redux';
 import Page from 'src/common/domain/Page';
 import { RootState } from 'src/common/presentation/state-module/root';
-import Notice from '../../domain/Notice';
+import Notice from '../../domain/model/Notice';
 import NoticeList from '../components/organisms/NoticeList';
 import * as listModule from "../state-module/list"
