+++ src/mother/notice/tests/infrastructure/model/Notice.unit.test.ts
@@ -1,4 +1,4 @@
-import Notice from "src/mother/notice/domain/Notice";
+import Notice from "src/mother/notice/domain/model/Notice";
 
 export const getNoticeFixture = (): Notice => ({
   "id": 1,
